Crate sallyport

API for the hypervisor-microkernel boundary

sallyport is a protocol crate for proxying service requests (such as syscalls) from an Enarx Keep to the host. A sally port is a secure gateway through which a defending army might “sally forth” from the protection of their fortification.

Mechanism of action

sallyport works by providing the host with the most minimal register context it requires to perform the syscall on the Keep’s behalf. In doing so, the host can immediately call the desired syscall without any additional logic required.

Guest and host side communicate via a mutually-distrusted shared block of memory.

This crate provides functionality for the guest to execute arbitary requests by means of using a Handler to proxy requests to the host via the untrusted block and corresponding functionality for the host to execute the requests contained within the untrusted block.

Phase-based request allocation

The crate identifies 3 distinct phases of an arbitrary call lifetime:

Stage

In this phase input references and output references are sequentially allocated within the untrusted sallyport block.

Once this phase is finished, no more allocations can be made within the untrusted sallyport block.

Entities that can be staged implement the Stage trait.

Commit

In this phase data is written to input references allocated in the stage phase. This may happen concurrently for N staged requests.

Once this phase is finished, the untrusted sallyport block is ready to be passed to the host for execution via platform-specific sally implementation.

Entities that can be committed implement the Commit trait.

Collect

In this phase data is read from output references allocated in the stage phase. This may happen concurrently for N staged requests.

This phase starts after the control returns to the guest. Once this phase is finished, the data within block is considered to be invalid, it may be left unchanged, but it may also be overwritten or dropped depending on request implementation

Entities that can be collected implement the Collect trait.

Modules

block

Block-specific functionality shared between guest and host.

guest

Guest-specific functionality.

host

Host-specific functionality.

Type Definitions

Error

Error type used within this crate.

Result

Result type returned by functionality exposed by this crate.