Trait sallyport::guest::alloc::PassthroughSyscall

pub unsafe trait PassthroughSyscall {
    type Argv: Into<[usize; 6]>;
    type Ret;

    const NUM: c_long;

    fn stage(self) -> Self::Argv;
}

Trait implemented by allocatable syscalls, which are passed through directly to the host and do not require custom handling logic.

Example

use sallyport::guest::syscall::Argv;

pub struct Exit {
    pub status: c_int,
}

unsafe impl PassthroughSyscall for Exit {
    const NUM: c_long = libc::SYS_exit;

    type Argv = Argv<1>;
    type Ret = ();

    fn stage(self) -> Self::Argv {
        Argv([self.status as _])
    }
}

Associated Types

type Argv: Into<[usize; 6]>

The syscall argument vector.

For example, syscall::Argv<1>.

type Ret

Syscall return value.

For example, ().

Associated Constants

const NUM: c_long

Syscall number.

For example, libc::SYS_exit.

Required methods

fn stage(self) -> Self::Argv

Returns argument vector registers.