Trait sallyport::guest::alloc::Syscall

pub unsafe trait Syscall<'a> {
    type Argv: Into<[usize; 6]>;
    type Ret: From<[usize; 2]>;
    type Staged: Commit<Item = Self::Committed>;
    type Committed;
    type Collected;

    const NUM: c_long;

    fn stage(
        self, 
        alloc: &mut impl Allocator
    ) -> Result<(Self::Argv, Self::Staged)>;

    fn collect(
        committed: Self::Committed, 
        ret: Self::Ret, 
        col: &impl Collector
    ) -> Self::Collected;
}

Trait implemented by allocatable syscalls.

Associated Types

type Argv: Into<[usize; 6]>

The syscall argument vector.

type Ret: From<[usize; 2]>

Syscall return value.

type Staged: Commit<Item = Self::Committed>

Opaque staged value value, which returns Self::Committed when committed via Commit::commit.

This is primarily designed to serve as a container for dynamic data allocated within stage.

type Committed

Opaque committed value returned by Commit::commit called upon Self::Staged, which is, in turn, passed to Self::collect to yield a Self::Collected.

type Collected

Value syscall collects as, which corresponds to its return value.

Associated Constants

const NUM: c_long

Syscall number.

Required methods

fn stage(self, alloc: &mut impl Allocator) -> Result<(Self::Argv, Self::Staged)>

Allocate dynamic data, if necessary and return resulting argument vector registers and opaque staged value on success.

fn collect(
committed: Self::Committed,
ret: Self::Ret,
col: &impl Collector
) -> Self::Collected

Collect the return registers and opaque committed value.