Trait sallyport::guest::call::alloc::Alloc

pub trait Alloc<'a, K> where
    K: Kind, {
    type Staged: Commit<Item = Self::Committed>;
    type Committed: Collect<Item = Self::Collected>;
    type Collected;
    fn stage(self, alloc: &mut impl Allocator) -> Result<Self::Staged>;
}

A generic call, which can be allocated within the block.

Associated Types

type Staged: Commit<Item = Self::Committed>

Opaque staged value, which returns Self::Committed when committed via Commit::commit.

This is designed to serve as a container for data allocated within stage.

type Committed: Collect<Item = Self::Collected>

Opaque committed value returned by Commit::commit called upon Self::Staged, which returns Self::Collected when collected via Collect::collect.

type Collected

Value call collects as. For example, a syscall return value.

Required methods

fn stage(self, alloc: &mut impl Allocator) -> Result<Self::Staged>

Allocate data, if necessary, and return resulting opaque staged value on success.