Trait sallyport::guest::call::enarxcall::PassthroughAlloc

pub trait PassthroughAlloc {
    type Argv: Into<[usize; 4]>;
    type Ret;

    const NUM: Number;

    fn stage(self) -> Self::Argv;
}

Trait implemented by allocatable Enarx calls, which are passed through directly to the host and do not require custom handling logic.

Associated Types

type Argv: Into<[usize; 4]>

The Enarx call argument vector.

For example, call::types::Argv<3>.

type Ret

Enarx call return value.

For example, usize.

Associated Constants

const NUM: Number

Enarx call number.

For example, Number::BalloonMemory.

Required methods

fn stage(self) -> Self::Argv

Returns argument vector registers.