Module sallyport::guest::call::gdbcall

GDB call-specific functionality.

Modules

types

Gdbcall-specific types.

Structs

Flush

gdbstub::conn::Connection::flush call

OnSessionStart

gdbstub::conn::Connection::on_session_start call

Peek

gdbstub::conn::ConnectionExt::peek call

Read

gdbstub::conn::ConnectionExt::read call

Write

gdbstub::conn::Connection::write call

WriteAll

gdbstub::conn::Connection::write_all call

Traits

Alloc

A generic GDB call, which can be allocated within the block.

PassthroughAlloc

Trait implemented by allocatable GDB calls, which are passed through directly to the host and do not require custom handling logic.