Trait sallyport::guest::call::gdbcall::Alloc

pub trait Alloc<'a> {
    type Argv: Into<[usize; 4]>;
    type Ret;
    type Staged: Commit<Item = Self::Committed>;
    type Committed;
    type Collected;

    const NUM: Number;

    fn stage(
        self, 
        alloc: &mut impl Allocator
    ) -> Result<(Self::Argv, Self::Staged)>;

    fn collect(
        committed: Self::Committed, 
        ret: Result<Self::Ret>, 
        col: &impl Collector
    ) -> Self::Collected;
}

A generic GDB call, which can be allocated within the block.

Examples

use sallyport::item::gdbcall::Number;
use sallyport::guest::alloc::{Allocator, Collector, Input};
use sallyport::guest::call::types::Argv;
use sallyport::guest::gdbcall::Alloc;
use sallyport::guest::gdbcall::types::{StagedBytesInput};
use sallyport::Result;

pub struct WriteAll<'a> {
    pub buf: &'a [u8],
}

impl<'a> Alloc<'a> for WriteAll<'a> {
    const NUM: Number = Number::WriteAll;

    type Argv = Argv<2>;
    type Ret = usize;

    type Staged = StagedBytesInput<'a>;
    type Committed = usize;
    type Collected = Option<Result<usize>>;

    fn stage(self, alloc: &mut impl Allocator) -> Result<(Self::Argv, Self::Staged)> {
        let (buf, _) = Input::stage_slice_max(alloc, self.buf)?;
        Ok((Argv([buf.offset(), buf.len()]), StagedBytesInput(buf)))
    }

    fn collect(
        count: Self::Committed,
        ret: Result<Self::Ret>,
        _: &impl Collector,
    ) -> Self::Collected {
        match ret {
            Ok(ret) if ret > count => None,
            res @ Ok(_) => Some(res),
            err => Some(err),
        }
    }
}

Associated Types

type Argv: Into<[usize; 4]>

The GDB call argument vector.

For example, guest::call::types::Argv<2>.

type Ret

GDB call return value.

For example, usize.

type Staged: Commit<Item = Self::Committed>

Opaque staged value, which returns Self::Committed when committed via Commit::commit.

This is designed to serve as a container for dynamic data allocated within stage.

For example, Input<'a, [u8], &'a [u8]>.

type Committed

Opaque committed value returned by guest::alloc::Commit::commit called upon Self::Staged, which returns Self::Collected when collected via guest::alloc::Collect::collect.

type Collected

Value GDB call collects as, which corresponds to its return value.

For example, Option<Result<usize>>.

Associated Constants

const NUM: Number

GDB call number.

For example, item::gdbcall::Number::WriteAll.

Required methods

fn stage(self, alloc: &mut impl Allocator) -> Result<(Self::Argv, Self::Staged)>

Allocate dynamic data, if necessary and return resulting argument vector registers and opaque staged value on success.

fn collect(
committed: Self::Committed,
ret: Result<Self::Ret>,
col: &impl Collector
) -> Self::Collected

Collect the return registers, opaque committed value and return a Self::Collected.