Trait sallyport::guest::call::gdbcall::PassthroughAlloc

pub trait PassthroughAlloc {
    type Argv: Into<[usize; 4]>;
    type Ret;

    const NUM: Number;

    fn stage(self) -> Self::Argv;
}

Trait implemented by allocatable GDB calls, which are passed through directly to the host and do not require custom handling logic.

Example

use sallyport::item::gdbcall::Number;
use sallyport::guest::call::types::Argv;
use sallyport::guest::gdbcall::PassthroughAlloc;
use sallyport::Result;

pub struct Read;

impl PassthroughAlloc for Read {
    const NUM: Number = Number::Read;

    type Argv = Argv<0>;
    type Ret = u8;

    fn stage(self) -> Self::Argv {
        Argv([])
    }
}

Associated Types

type Argv: Into<[usize; 4]>

The gdbcall argument vector.

For example, call::types::Argv<0>.

type Ret

Gdbcall return value.

For example, u8.

Associated Constants

const NUM: Number

GDB call number.

For example, Number::Read.

Required methods

fn stage(self) -> Self::Argv

Returns argument vector registers.