Module sallyport::guest::call::syscall

Syscall-specific functionality.

Modules

types

Syscall-specific types.

Structs

Accept

Accept4

AllocIoctl

AllocOpen

Bind

Close

Connect

Dup

Dup2

Dup3

EpollCreate1

EpollCtl

EpollPwait

EpollWait

Eventfd2

Exit

ExitGroup

Fcntl

Fstat

Getegid

Geteuid

Getgid

Getpid

Getrandom

Getsockname

Getuid

Ioctl

Listen

Nanosleep

Open

Poll

Read

Readlink

Readv

Recv

Recvfrom

RtSigprocmask

Send

Sendto

SetTidAddress

Setsockopt

Sigaltstack

Socket

StagedNanosleep

StagedSendto

StagedSetsockopt

Sync

Uname

Write

Writev

Constants

FAKE_GID

Fake GID returned by enarx.

FAKE_PID

Fake PID returned by enarx.

FAKE_TID

Fake TID returned by enarx.

FAKE_UID

Fake UID returned by enarx.

Traits

Alloc

A generic syscall, which can be allocated within the block.

PassthroughAlloc

Trait implemented by allocatable syscalls, which are passed through directly to the host and do not require custom handling logic.