Trait sallyport::guest::call::syscall::Alloc

pub unsafe trait Alloc<'a> {
    type Argv: Into<[usize; 6]>;
    type Ret;
    type Staged: Commit<Item = Self::Committed>;
    type Committed;
    type Collected;

    const NUM: c_long;

    fn stage(
        self, 
        alloc: &mut impl Allocator
    ) -> Result<(Self::Argv, Self::Staged)>;

    fn collect(
        committed: Self::Committed, 
        ret: Result<Self::Ret>, 
        col: &impl Collector
    ) -> Self::Collected;
}

A generic syscall, which can be allocated within the block.

Safety

This trait is unsafe, because it allows execution arbitrary syscalls on the host, which is intrinsically unsafe.

Examples

use sallyport::guest::alloc::{Allocator, Collector, Output};
use sallyport::guest::call::types::Argv;
use sallyport::guest::syscall::Alloc;
use sallyport::Result;

pub struct Read<'a> {
   pub fd: c_int,
   pub buf: &'a mut [u8],
}

unsafe impl<'a> Alloc<'a> for Read<'a> {
    const NUM: c_long = libc::SYS_read;

    type Argv = Argv<3>;
    type Ret = size_t;

    type Staged = Output<'a, [u8], &'a mut [u8]>;
    type Committed = Self::Staged;
    type Collected = Option<Result<size_t>>;

    fn stage(self, alloc: &mut impl Allocator) -> Result<(Self::Argv, Self::Staged)> {
        let (buf, _) = Output::stage_slice_max(alloc, self.buf)?;
        Ok((Argv([self.fd as _, buf.offset(), buf.len()]), buf))
    }

    fn collect(
        buf: Self::Committed,
        ret: Result<Self::Ret>,
        col: &impl Collector,
    ) -> Self::Collected {
        match ret {
            Ok(ret) if ret > buf.len() => None,
            res @ Ok(ret) => {
                unsafe { buf.collect_range(col, 0..ret) };
                Some(res)
            }
            err => Some(err),
        }
    }
}

Associated Types

type Argv: Into<[usize; 6]>

The syscall argument vector.

For example, guest::call::types::Argv<3>.

type Ret

Syscall return value.

For example, libc::size_t.

type Staged: Commit<Item = Self::Committed>

Opaque staged value, which returns Self::Committed when committed via Commit::commit.

This is designed to serve as a container for dynamic data allocated within stage.

For example, Output<'a, [u8], &'a mut [u8]>.

type Committed

Opaque committed value returned by guest::alloc::Commit::commit called upon Self::Staged, which returns Self::Collected when collected via guest::alloc::Collect::collect.

type Collected

Value syscall collects as, which corresponds to its return value.

For example, Option<Result<libc::size_t>>.

Associated Constants

const NUM: c_long

Syscall number.

For example, libc::SYS_read.

Required methods

fn stage(self, alloc: &mut impl Allocator) -> Result<(Self::Argv, Self::Staged)>

Allocate dynamic data, if necessary and return resulting argument vector registers and opaque staged value on success.

fn collect(
committed: Self::Committed,
ret: Result<Self::Ret>,
col: &impl Collector
) -> Self::Collected

Collect the return registers, opaque committed value and return a Self::Collected.