Trait sallyport::guest::call::syscall::PassthroughAlloc

pub unsafe trait PassthroughAlloc {
    type Argv: Into<[usize; 6]>;
    type Ret;

    const NUM: c_long;

    fn stage(self) -> Self::Argv;
}

Trait implemented by allocatable syscalls, which are passed through directly to the host and do not require custom handling logic.

Safety

This trait is unsafe, because it allows execution arbitrary syscalls on the host, which is intrinsically unsafe.

Example

use sallyport::guest::call::types::Argv;
use sallyport::guest::syscall::PassthroughAlloc;
use sallyport::Result;

pub struct Exit {
    pub status: c_int,
}

unsafe impl PassthroughAlloc for Exit {
    const NUM: c_long = libc::SYS_exit;

    type Argv = Argv<1>;
    type Ret = ();

    fn stage(self) -> Self::Argv {
        Argv([self.status as _])
    }
}

Associated Types

type Argv: Into<[usize; 6]>

The syscall argument vector.

For example, call::types::Argv<1>.

type Ret

Syscall return value.

For example, ().

Associated Constants

const NUM: c_long

Syscall number.

For example, [libc::SYS_exit].

Required methods

fn stage(self) -> Self::Argv

Returns argument vector registers.