Trait sallyport::guest::Platform

pub trait Platform {
    fn sally(&mut self) -> Result<(), c_int>;

    fn validate_mut<'a, T>(&self, ptr: usize) -> Result<&'a mut T, c_int>;

    fn validate_slice_mut<'a, T>(
        &self, 
        ptr: usize, 
        len: usize
    ) -> Result<&'a mut [T], c_int>;

    fn validate<'a, T>(&self, ptr: usize) -> Result<&'a T, c_int> { ... }

    fn validate_slice<'a, T>(
        &self, 
        ptr: usize, 
        len: usize
    ) -> Result<&'a [T], c_int> { ... }
}

Platform-specific functionality.

Required methods

fn sally(&mut self) -> Result<(), c_int>

Suspend guest execution and pass control to host. This function will return when the host passes control back to the guest.

fn validate_mut<'a, T>(&self, ptr: usize) -> Result<&'a mut T, c_int>

Validates that a region of memory is valid for read-write access. Returns a mutable borrow if valid, otherwise EINVAL.

fn validate_slice_mut<'a, T>(
&self,
ptr: usize,
len: usize
) -> Result<&'a mut [T], c_int>

Validates that a region of memory is valid for read-write access for len elements of type T. Returns a mutable borrow if valid, otherwise EINVAL.

Provided methods

fn validate<'a, T>(&self, ptr: usize) -> Result<&'a T, c_int>

Validates that a region of memory is valid for read-only access. Returns an immutable borrow if valid, otherwise EINVAL.

fn validate_slice<'a, T>(
&self,
ptr: usize,
len: usize
) -> Result<&'a [T], c_int>

Validates that a region of memory is valid for read-only access for len elements of type T. Returns a mutable borrow if valid, otherwise EINVAL.