Trait sallyport::iter::Iterator

pub trait Iterator: Sized {
    type Item;
    fn next(self) -> Option<(Self::Item, Self)>;

    fn fold<T, F>(self, init: T, f: F) -> T
    where
        F: FnMut(T, Self::Item) -> T,
    { ... }

    fn for_each<F>(self, f: F)
    where
        Self: Sized,
        F: FnMut(Self::Item),
    { ... }
}

An interface for dealing with consuming, tail-recursive iterators.

This is an adaptation of core::iter::Iterator. For more about the concept of iterators generally, please see the core::iter.

Associated Types

type Item

Required methods

fn next(self) -> Option<(Self::Item, Self)>

Consumes the iterator and returns a tuple of the next value and iterator tail.

Returns None when iteration is finished.

Examples

Basic usage:

use sallyport::iter::Iterator;

let a = [1, 2, 3];

let iter = a.iter();

// A call to next() returns the next value...
let (item, iter) = Iterator::next(iter).unwrap();
assert_eq!(&1, item);
let (item, iter) = iter.next().unwrap();
assert_eq!(&2, item);
let (item, iter) = iter.next().unwrap();
assert_eq!(&3, item);

// ... and then None once it's over.
assert!(iter.next().is_none());

Provided methods

fn fold<T, F>(self, init: T, f: F) -> T where
F: FnMut(T, Self::Item) -> T,

fn for_each<F>(self, f: F) where
Self: Sized,
F: FnMut(Self::Item),