Trait sallyport::syscall::ProcessSyscallHandler

pub trait ProcessSyscallHandler: BaseSyscallHandler + AddressValidator + Sized {
    fn arch_prctl(&mut self, code: c_int, addr: c_ulong) -> Result;

    fn exit(&mut self, status: c_int) -> ! { ... }

    fn exit_group(&mut self, status: c_int) -> ! { ... }

    fn set_tid_address(&mut self, _tidptr: *const c_int) -> Result { ... }

    fn rt_sigaction(
        &mut self, 
        signum: c_int, 
        act: UntrustedRef<'_, KernelSigAction>, 
        oldact: UntrustedRefMut<'_, KernelSigAction>, 
        size: usize
    ) -> Result { ... }

    fn rt_sigprocmask(
        &mut self, 
        _how: c_int, 
        _set: UntrustedRef<'_, KernelSigSet>, 
        _oldset: UntrustedRefMut<'_, KernelSigSet>, 
        _sigsetsize: size_t
    ) -> Result { ... }

    fn sigaltstack(
        &mut self, 
        _ss: UntrustedRef<'_, stack_t>, 
        _old_ss: UntrustedRefMut<'_, stack_t>
    ) -> Result { ... }

    fn getpid(&mut self) -> Result { ... }

    fn getuid(&mut self) -> Result { ... }

    fn getgid(&mut self) -> Result { ... }

    fn geteuid(&mut self) -> Result { ... }

    fn getegid(&mut self) -> Result { ... }
}

process syscalls

Required methods

fn arch_prctl(&mut self, code: c_int, addr: c_ulong) -> Result

syscall

Provided methods

fn exit(&mut self, status: c_int) -> !

Proxy an exit() syscall

fn exit_group(&mut self, status: c_int) -> !

Proxy an exitgroup() syscall

TODO: Currently we are only using one thread, so this will behave the same way as exit(). In the future, this implementation will change.

fn set_tid_address(&mut self, _tidptr: *const c_int) -> Result

Do a set_tid_address() syscall

This is currently unimplemented and returns a dummy thread id.

fn rt_sigaction(
&mut self,
signum: c_int,
act: UntrustedRef<'_, KernelSigAction>,
oldact: UntrustedRefMut<'_, KernelSigAction>,
size: usize
) -> Result

Do a rt_sigaction() system call

We don’t support signals yet. So, fake success.

fn rt_sigprocmask(
&mut self,
_how: c_int,
_set: UntrustedRef<'_, KernelSigSet>,
_oldset: UntrustedRefMut<'_, KernelSigSet>,
_sigsetsize: size_t
) -> Result

Do a rt_sigprocmask() syscall

We don’t support signals yet. So, fake success.

fn sigaltstack(
&mut self,
_ss: UntrustedRef<'_, stack_t>,
_old_ss: UntrustedRefMut<'_, stack_t>
) -> Result

Do a sigaltstack() syscall

This is currently unimplemented and returns success.

fn getpid(&mut self) -> Result

syscall

fn getuid(&mut self) -> Result

Do a getuid() syscall

fn getgid(&mut self) -> Result

Do a getgid() syscall

fn geteuid(&mut self) -> Result

Do a geteuid() syscall

fn getegid(&mut self) -> Result

Do a getegid() syscall