Trait sallyport::guest::Execute

pub trait Execute {
    type Platform: Platform;
    type Allocator: Allocator;
    fn platform(&mut self) -> &mut Self::Platform;

    fn allocator(&mut self) -> Self::Allocator;

    fn execute<'a, T>(
        &mut self, 
        req: impl Stage<'a, Item = impl Commit<Item = impl Collect<Item = T>>>
    ) -> Result<T> { ... }

    fn attacked(&mut self) -> ! { ... }

    unsafe fn syscall(&mut self, registers: [usize; 7]) -> Result<[usize; 2]> { ... }

    fn exit(&mut self, status: c_int) -> Result<()> { ... }

    fn read(&mut self, fd: c_int, buf: &mut [u8]) -> Result<size_t> { ... }

    fn write(&mut self, fd: c_int, buf: &[u8]) -> Result<size_t> { ... }
}

Associated Types

type Platform: Platform

type Allocator: Allocator

Required methods

fn platform(&mut self) -> &mut Self::Platform

fn allocator(&mut self) -> Self::Allocator

Provided methods

fn execute<'a, T>(
&mut self,
req: impl Stage<'a, Item = impl Commit<Item = impl Collect<Item = T>>>
) -> Result<T>

Executes an arbitrary call. Examples of calls that this method can execute are:

syscall::Read

syscall::Exit

fn attacked(&mut self) -> !

Loops infinitely trying to exit.

unsafe fn syscall(&mut self, registers: [usize; 7]) -> Result<[usize; 2]>

Executes a supported syscall expressed as an opaque 7-word array akin to libc::syscall.

fn exit(&mut self, status: c_int) -> Result<()>

Executes exit syscall akin to libc::exit.

fn read(&mut self, fd: c_int, buf: &mut [u8]) -> Result<size_t>

Executes read syscall akin to libc::read.

fn write(&mut self, fd: c_int, buf: &[u8]) -> Result<size_t>

Executes write syscall akin to libc::write.