Trait sallyport::guest::Execute

pub trait Execute {

    fn execute<'a, T: Call<'a>>(&mut self, call: T) -> Result<T::Collected>;

    unsafe fn syscall(&mut self, registers: [usize; 7]) -> Result<[usize; 2]>;

    fn attacked(&mut self) -> ! { ... }

    fn bind(&mut self, sockfd: c_int, addr: &[u8]) -> Result<()> { ... }

    fn clock_gettime(
        &mut self, 
        clockid: clockid_t, 
        tp: &mut timespec
    ) -> Result<()> { ... }

    fn close(&mut self, fd: c_int) -> Result<()> { ... }

    fn connect(&mut self, sockfd: c_int, addr: &[u8]) -> Result<()> { ... }

    fn dup(&mut self, oldfd: c_int) -> Result<()> { ... }

    fn dup2(&mut self, oldfd: c_int, newfd: c_int) -> Result<()> { ... }

    fn dup3(&mut self, oldfd: c_int, newfd: c_int, flags: c_int) -> Result<()> { ... }

    fn eventfd2(&mut self, initval: c_int, flags: c_int) -> Result<c_int> { ... }

    fn exit(&mut self, status: c_int) -> Result<()> { ... }

    fn exit_group(&mut self, status: c_int) -> Result<()> { ... }

    fn fcntl(&mut self, fd: c_int, cmd: c_int, arg: c_int) -> Result<c_int> { ... }

    fn fstat(&mut self, fd: c_int, statbuf: &mut stat) -> Result<()> { ... }

    fn getegid(&mut self) -> Result<gid_t> { ... }

    fn geteuid(&mut self) -> Result<uid_t> { ... }

    fn getgid(&mut self) -> Result<gid_t> { ... }

    fn getpid(&mut self) -> Result<pid_t> { ... }

    fn getrandom(&mut self, buf: &mut [u8], flags: c_uint) -> Result<size_t> { ... }

    fn getsockname(
        &mut self, 
        sockfd: c_int, 
        addr: SockaddrOutput<'_>
    ) -> Result<()> { ... }

    fn getuid(&mut self) -> Result<uid_t> { ... }

    fn listen(&mut self, sockfd: c_int, backlog: c_int) -> Result<()> { ... }

    fn read(&mut self, fd: c_int, buf: &mut [u8]) -> Result<size_t> { ... }

    fn setsockopt(
        &mut self, 
        sockfd: c_int, 
        level: c_int, 
        optname: c_int, 
        optval: &[u8]
    ) -> Result<c_int> { ... }

    fn socket(
        &mut self, 
        domain: c_int, 
        typ: c_int, 
        protocol: c_int
    ) -> Result<c_int> { ... }

    fn sync(&mut self) -> Result<()> { ... }

    fn uname(&mut self, buf: &mut utsname) -> Result<()> { ... }

    fn write(&mut self, fd: c_int, buf: &[u8]) -> Result<size_t> { ... }

}

Required methods

fn execute<'a, T: Call<'a>>(&mut self, call: T) -> Result<T::Collected>

Executes an arbitrary call. Examples of calls that this method can execute are:

syscall::Exit

syscall::Read

syscall::Write

unsafe fn syscall(&mut self, registers: [usize; 7]) -> Result<[usize; 2]>

Executes a supported syscall expressed as an opaque 7-word array akin to libc::syscall.

Provided methods

fn attacked(&mut self) -> !

Loops infinitely trying to exit.

fn bind(&mut self, sockfd: c_int, addr: &[u8]) -> Result<()>

Executes bind syscall akin to libc::bind.

fn clock_gettime(&mut self, clockid: clockid_t, tp: &mut timespec) -> Result<()>

Executes clock_gettime syscall akin to libc::clock_gettime.

fn close(&mut self, fd: c_int) -> Result<()>

Executes close syscall akin to libc::close.

fn connect(&mut self, sockfd: c_int, addr: &[u8]) -> Result<()>

Executes connect syscall akin to libc::connect.

fn dup(&mut self, oldfd: c_int) -> Result<()>

Executes dup syscall akin to libc::dup.

fn dup2(&mut self, oldfd: c_int, newfd: c_int) -> Result<()>

Executes dup2 syscall akin to libc::dup2.

fn dup3(&mut self, oldfd: c_int, newfd: c_int, flags: c_int) -> Result<()>

Executes dup3 syscall akin to libc::dup3.

fn eventfd2(&mut self, initval: c_int, flags: c_int) -> Result<c_int>

Executes eventfd2.

fn exit(&mut self, status: c_int) -> Result<()>

Executes exit syscall akin to libc::exit.

fn exit_group(&mut self, status: c_int) -> Result<()>

Executes exit_group.

fn fcntl(&mut self, fd: c_int, cmd: c_int, arg: c_int) -> Result<c_int>

Executes fcntl syscall akin to libc::fcntl.

fn fstat(&mut self, fd: c_int, statbuf: &mut stat) -> Result<()>

Executes fstat syscall akin to libc::fstat.

fn getegid(&mut self) -> Result<gid_t>

Executes getegid syscall akin to libc::getegid.

fn geteuid(&mut self) -> Result<uid_t>

Executes geteuid syscall akin to libc::geteuid.

fn getgid(&mut self) -> Result<gid_t>

Executes getgid syscall akin to libc::getgid.

fn getpid(&mut self) -> Result<pid_t>

Executes getpid syscall akin to libc::getpid.

fn getrandom(&mut self, buf: &mut [u8], flags: c_uint) -> Result<size_t>

Executes getrandom syscall akin to libc::getrandom.

fn getsockname(&mut self, sockfd: c_int, addr: SockaddrOutput<'_>) -> Result<()>

Executes getsockname syscall akin to libc::getsockname.

fn getuid(&mut self) -> Result<uid_t>

Executes getuid syscall akin to libc::getuid.

fn listen(&mut self, sockfd: c_int, backlog: c_int) -> Result<()>

Executes listen syscall akin to libc::listen.

fn read(&mut self, fd: c_int, buf: &mut [u8]) -> Result<size_t>

Executes read syscall akin to libc::read.

fn setsockopt(
&mut self,
sockfd: c_int,
level: c_int,
optname: c_int,
optval: &[u8]
) -> Result<c_int>

Executes setsockopt syscall akin to libc::setsockopt.

fn socket(
&mut self,
domain: c_int,
typ: c_int,
protocol: c_int
) -> Result<c_int>

Executes socket syscall akin to libc::socket.

fn sync(&mut self) -> Result<()>

Executes sync syscall akin to libc::sync.

fn uname(&mut self, buf: &mut utsname) -> Result<()>

Executes uname syscall akin to libc::uname.

fn write(&mut self, fd: c_int, buf: &[u8]) -> Result<size_t>

Executes write syscall akin to libc::write.