Trait sallyport::guest::Platform

pub trait Platform {
    fn validate_mut<T>(&self, ptr: usize) -> Result<&mut T, c_int>;

    fn validate<T>(&self, ptr: usize) -> Result<&T, c_int>;

    fn validate_slice_mut<T: Sized>(
        &self, 
        ptr: usize, 
        count: usize
    ) -> Result<&mut [T], c_int>;

    fn validate_slice<T: Sized>(
        &self, 
        ptr: usize, 
        count: usize
    ) -> Result<&[T], c_int>;

    fn validate_iovec_slice_mut(
        &self, 
        iov: usize, 
        iovcnt: usize
    ) -> Result<&mut [&mut [u8]], c_int> { ... }

    fn validate_iovec_slice(
        &self, 
        iov: usize, 
        iovcnt: usize
    ) -> Result<&[&[u8]], c_int> { ... }

    fn validate_str(&self, ptr: usize) -> Result<&[u8], c_int> { ... }

    fn validate_sockaddr_output(
        &self, 
        addr: usize, 
        addrlen: usize
    ) -> Result<SockaddrOutput<'_>, c_int> { ... }
}

Platform-specific functionality.

Required methods

fn validate_mut<T>(&self, ptr: usize) -> Result<&mut T, c_int>

Validates that the memory pointed to by ptr of the type T is:

in valid address space and writable for the lifetime of self.

“dereferenceable” in the sense defined in the ptr module documentation.

ptr is non-null and aligned

not borrowed already and registers the memory as borrowed mutably.

Returns a mutable borrow if valid, otherwise EINVAL.

fn validate<T>(&self, ptr: usize) -> Result<&T, c_int>

Validates that the memory pointed to by ptr of the type T is:

in valid address space and readable for the lifetime of self.

“dereferenceable” in the sense defined in the ptr module documentation.

ptr is non-null and aligned

not borrowed already and registers the memory as borrowed.

Returns an immutable borrow if valid, otherwise EINVAL.

fn validate_slice_mut<T: Sized>(
&self,
ptr: usize,
count: usize
) -> Result<&mut [T], c_int>

Validates that a region for count elements of type T is:

in valid address space and writable for the lifetime of self.

“dereferenceable” in the sense defined in the ptr module documentation.

ptr is non-null and aligned

not borrowed already and registers the memory as borrowed mutably.

Returns a mutable borrow if valid, otherwise EINVAL.

fn validate_slice<T: Sized>(
&self,
ptr: usize,
count: usize
) -> Result<&[T], c_int>

Validates that a region of memory is valid for read-only access for count elements of type T.

in valid address space and writable for the lifetime of self.

“dereferenceable” in the sense defined in the ptr module documentation.

ptr is non-null and aligned

not borrowed already and registers the memory as borrowed.

Returns an immutable borrow if valid, otherwise EINVAL.

Provided methods

fn validate_iovec_slice_mut(
&self,
iov: usize,
iovcnt: usize
) -> Result<&mut [&mut [u8]], c_int>

Validates that pointer iov points to represents a slice of iovcnt pointers to libc::iovec structures valid for read-write access.

Also checks, that memory is:

in valid address space and writable for the lifetime of self.

“dereferenceable” in the sense defined in [the ptr module documentation].

not borrowed already

and pointers are non-null and aligned and registers the memory as borrowed mutably.

Returns a mutable borrow of the libc::iovec structs as a 2-dimensional slice of mutable byte buffer borrows if valid, otherwise EINVAL.

fn validate_iovec_slice(
&self,
iov: usize,
iovcnt: usize
) -> Result<&[&[u8]], c_int>

Validates that pointer iov points to represents a slice of iovcnt pointers to libc::iovec structures valid for read-only access.

Also checks, that the memory is:

in valid address space and readable for the lifetime of self.

“dereferenceable” in the sense defined in [the ptr module documentation].

not borrowed already

and pointers are non-null and aligned and registers the memory as borrowed.

Returns an immutable borrow of the libc::iovec structs as a 2-dimensional slice of immutable byte buffer borrows if valid, otherwise EINVAL.

fn validate_str(&self, ptr: usize) -> Result<&[u8], c_int>

Validates that a region of memory represents a C string and is valid for read-only access.

Also checks, that the memory is:

in valid address space and readable for the lifetime of self.

“dereferenceable” in the sense defined in [the ptr module documentation].

ptr is non-null and aligned

not borrowed already and registers the memory as borrowed.

Returns an immutable borrow of bytes of the string with the nul terminator byte if valid, otherwise EINVAL.

fn validate_sockaddr_output(
&self,
addr: usize,
addrlen: usize
) -> Result<SockaddrOutput<'_>, c_int>

Validates that pointer addrlen points to socklen_t and addr points to a byte array of size *addrlen and is valid for read-write access.

Also checks that the memory is:

in valid address space and writable for the lifetime of self.

“dereferenceable” in the sense defined in [the ptr module documentation].

not borrowed already

and pointers are non-null and aligned and registers the memory as borrowed.

Returns a SockaddrOutput, otherwise EINVAL.